Microaneurysm Detection for Screening of Diabetic Retinopathy
Microaneurosysms are the initial signs of diabetic
retinopathy which appears as small circular reddish dots in the
retina. For automatic analysis of diabetic retinopathy, reliable
detection of microaneurysms is an important task. An analytical
approach for detecting automated microaneurysm accurately is
proposed in this work. Dark object filtering is applied initially
for locating candidate objects in the preprocessed retinal fundus
image. Then blood vessels are removed from the preprocessed
image followed by identification of circular structured regions
in the retinal image, that facilitates the extraction of microaneurysms.
This leads to increase in the profile differences between
true microaneurysms and other non-microaneurysm candidates.
These profiles are subjected to statistical feature extraction and
classified with the help of K-star classifier. Microaneurysms were
well separated from the retinal background by this method. This
approach demonstrated robustness on large scale dataset testing
with specificity and sensitivity accepted clinically. This system
can be a powerful screening tool for diabetic retinopathy..